Aula 22/08 - Término da API, com paginação e esqueleto da documentação

+++ index.py
@@ -1,27 +1,34 @@
+from datetime import datetime
 from http.client import BAD_REQUEST, NOT_FOUND, UNAUTHORIZED, UNPROCESSABLE_ENTITY
 import json
-from typing import List
 import logging
 from flask import (
     Flask,
     session,
-    escape,
-    render_template,
     request,
 )
 from config import setup_database
 
-from model import Contato, Usuario, db
+from model import Contato, ContatosPaginatedSchema, Usuario, db
 
 from flask_migrate import Migrate
 
 from flask_session import Session
+from utils import pagination_to_json
 
 from validacao_input import validar_json
 
 from flask_cors import CORS
 
+from flask_sqlalchemy import Pagination
+
+from flask_pydantic_spec import FlaskPydanticSpec, Response, Request
+
+
 app = Flask(__name__)
+
+spec = FlaskPydanticSpec("flask", title="Contatos API")
+spec.register(app)
 
 app.config["SESSION_PERMANENT"] = False
 app.config["SESSION_TYPE"] = "filesystem"
@@ -39,71 +46,92 @@ CORS(app, supports_credentials=True, resources={r"/*": {"origins": "*"}})  # TOD
 logging.getLogger('flask_cors').level = logging.DEBUG
 
 
-@app.get("/contatos")
-def contatos_json():
-
-     # Pegar usuario logado, se existir:
+def verificar_login():
+    # Pegar usuario logado, se existir:
     usuario = None
     if "user" in session.keys():
         usuario = session["user"]
 
     if not usuario:
+        raise ValueError   ## TODO: Criar exceção especifica
+
+    return usuario
+
+@app.get("/contatos")
+@spec.validate(resp=Response(HTTP_200=ContatosPaginatedSchema)
+def contatos_json():
+
+    try:
+        usuario = verificar_login()
+    except:
         return {"erro": "Usuário não logado"}, UNAUTHORIZED
 
     id_usuario = usuario.id
+
+    if request.args.get("page"):
+        page = int(request.args.get("page"))  ## pode dar erro
+    else:
+        page = 1
 
     if request.args.get("nome"):
         busca_str = "%" + request.args.get("nome") + "%"
-        contatos:List[Contato] = (
+        contatos:Pagination = (
             Contato.query                                # objeto Query
                     .filter_by(id_usuario=id_usuario)    # objeto Query
                     .filter(Contato.nome.ilike(busca_str))  # objeto Query
-                    .all()  # List[Contato]
+                    .paginate(page=page, per_page=PER_PAGE)  # Pagination
         )
     else:
     
-        contatos:List[Contato] = (
+        contatos:Pagination = (
             Contato.query                                # objeto Query
                     .filter_by(id_usuario=id_usuario)    # objeto Query
-                    .all()  # List[Contato]
+                    .paginate(page=page, per_page=PER_PAGE)  # Pagination
+
         )
 
-    if not contatos:
-        return [], NOT_FOUND
 
-    return json.dumps([contato.as_dict() for contato in contatos])
+    contatos_json = pagination_to_json(contatos, "http://localhost:5000")
+
+    return contatos_json
 
 
-@app.get("/contatos_fixo")
-def contatos_json_sem_login():
+# @app.get("/contatos_fixo")
+# def contatos_json_sem_login():
 
-     # Pegar usuario logado, se existir:
-    usuario = None
-    if "user" in session.keys():
-        usuario = session["user"]
+#      # Pegar usuario logado, se existir:
+#     usuario = None
+#     if "user" in session.keys():
+#         usuario = session["user"]
 
-    if not usuario:
-        return {"erro": "Usuário não logado"}, UNAUTHORIZED
+#     if not usuario:
+#         return {"erro": "Usuário não logado"}, UNAUTHORIZED
 
-    id_usuario = usuario.id
+#     id_usuario = usuario.id
     
-    contatos:List[Contato] = (
-        Contato.query                                # objeto Query
-                .filter_by(id_usuario=id_usuario)    # objeto Query
-                # .filter(Contato.nome.ilike(busca_str))  # objeto Query
-                .all()  # List[Contato]
-    )
+#     contatos:List[Contato] = (
+#         Contato.query                                # objeto Query
+#                 .filter_by(id_usuario=id_usuario)    # objeto Query
+#                 # .filter(Contato.nome.ilike(busca_str))  # objeto Query
+#                 .all()  # List[Contato]
+#     )
 
-    if not contatos:
-        return [], NOT_FOUND
+#     if not contatos:
+#         return [], NOT_FOUND
 
-    return json.dumps([contato.as_dict() for contato in contatos])
+#     return json.dumps([contato.to_json() for contato in contatos])
 
 
 @app.get("/contatos/<id_>")
 def contato(id_):
 
-    id_usuario = 2   ## TODO: depois trocar pela validação de sessão
+    try:
+        usuario = verificar_login()
+    except:
+        return {"erro": "Usuário não logado"}, UNAUTHORIZED
+
+    id_usuario = usuario.id
+
 
     contato = (
         Contato.query                                # objeto Query
@@ -118,7 +146,7 @@ def contato(id_):
         # resp.json({"erro": "Usuario não existe"})
         return {"erro": "Usuario não existe"}, NOT_FOUND
 
-    return json.dumps(contato.as_dict())
+    return json.dumps(contato.to_json())
 
 
 
@@ -127,6 +155,14 @@ def contato(id_):
 
 @app.post("/contatos")
 def adicionar_contato_action():
+
+    try:
+        usuario = verificar_login()
+    except:
+        return {"erro": "Usuário não logado"}, UNAUTHORIZED
+
+    id_usuario = usuario.id
+
 
     resultado_validacao = validar_json(request.json)
 
@@ -149,37 +185,61 @@ def adicionar_contato_action():
         telefone=telefone, 
         data_nascimento=data_nascimento,
         detalhes=detalhes,
-        id_usuario=2)
+        id_usuario=id_usuario)
 
     db.session.add(contato)  # adiciona ou atualiza
     db.session.commit()
 
     ## TODO: SETAR ID NO RETORNO
     
-    return json.dumps(contato.as_dict())
+    return json.dumps(contato.to_json())
 
 
 @app.delete("/contatos/<id_>")
 def remover_contato_action(id_):
 
+    try:
+        usuario = verificar_login()
+    except:
+        return {"erro": "Usuário não logado"}, UNAUTHORIZED
+
+    id_usuario = usuario.id
+
     contato = Contato.query.filter_by(id=id_).first()
 
     if not contato:
-        return {"erro": "Usuário não existe"}
+        return {"erro": "Usuário não existe"}, NOT_FOUND
+
+    if contato.id_usuario != id_usuario:
+        return {"erro": "Contato não pertence ao usuário"}, UNAUTHORIZED
+
 
     db.session.delete(contato)
     db.session.commit()
 
-    return json.dumps(contato.as_dict())
+    return json.dumps(contato.to_json())
 
 
 @app.put("/contatos/<id_>")
 def put_contato(id_):
 
+     # Pegar usuario logado, se existir:
+    usuario = None
+    if "user" in session.keys():
+        usuario = session["user"]
+
+    if not usuario:
+        return {"erro": "Usuário não logado"}, UNAUTHORIZED
+
+    id_usuario = usuario.id
+
     contato = Contato.query.filter_by(id=id_).first()
 
     if not contato:
         return {"erro": "Usuario nao existe"}, UNPROCESSABLE_ENTITY
+
+    if contato.id_usuario != id_usuario:
+        return {"erro": "Contato não pertence ao usuário"}, UNAUTHORIZED
 
     # id_ = request.json["id"]
     nome = request.json["nome"]
@@ -198,16 +258,27 @@ def put_contato(id_):
     db.session.merge(contato)  # adiciona ou atualiza
     db.session.commit()
 
-    return json.dumps(contato.as_dict())
+    return json.dumps(contato.to_json())
 
     
 
 @app.patch("/contatos/<id_>")
 def patch_contato(id_):
+
+    try:
+        usuario = verificar_login()
+    except:
+        return {"erro": "Usuário não logado"}, UNAUTHORIZED
+
+    id_usuario = usuario.id
+
     contato = Contato.query.filter_by(id=id_).first()
 
     if not contato:
         return {"erro": "Usuario nao existe"}, UNPROCESSABLE_ENTITY
+
+    if contato.id_usuario != id_usuario:
+        return {"erro": "Contato não pertence ao usuário"}, UNAUTHORIZED
 
     for nome_campo, valor in request.json.items():
         # ex: "nome"  -> "Marcos Tacalepau Nesse Carrinho"
@@ -219,11 +290,12 @@ def patch_contato(id_):
     db.session.merge(contato)  # adiciona ou atualiza
     db.session.commit()
 
-    return json.dumps(contato.as_dict())
+    return json.dumps(contato.to_json())
 
 
 @app.post("/usuarios")
 def cadastrar_usuario_action():
+
 
     username = request.json.get("username")
     senha = request.json.get("senha")
@@ -242,7 +314,7 @@ def cadastrar_usuario_action():
 
     db.session.commit() ## COMMIT DA TRANSAÇÃO
 
-    dict_usuario = usuario.as_dict()
+    dict_usuario = usuario.to_json()
     del dict_usuario["password_hash"]
 
     return json.dumps(dict_usuario)
@@ -266,7 +338,7 @@ def login_action():
     # else
     session["user"] = usuario
 
-    dict_usuario = usuario.as_dict()
+    dict_usuario = usuario.to_json()
     del dict_usuario["password_hash"]
 
     return json.dumps(dict_usuario)
@@ -278,20 +350,7 @@ def logout_action(id_):
     usuario = session.get("user")
     session.pop("user")
 
-    dict_usuario = usuario.as_dict()
+    dict_usuario = usuario.to_json()
     del dict_usuario["password_hash"]
 
     return json.dumps(dict_usuario)
-
-
-@app.get("/form_test_xss")
-def form_test_xss():
-    return render_template("form_test.html")
-
-
-@app.post("/form_test_action")
-def form_test_action():
-    campo = request.form["campo"]
-
-    return "<p> {} </p>".format(escape(campo))
-
